Restrict application routes to admin only

+++ server/api/controllers/ResponseController.ts
@@ -1,6 +1,7 @@
 import {
   BodyParam,
   Delete,
+  ForbiddenError,
   Get,
   JsonController,
   Patch,
@@ -22,6 +23,7 @@ import { Application } from '../validators/ResponseRequests';
 import { StorageService } from '../../services/StorageService';
 import { MediaType } from '../../types/Enums';
 import { File } from '../../types/ApiRequests';
+import PermissionsService from '../../services/PermissionsService';
 
 @JsonController('/response')
 @Service()
@@ -60,6 +62,9 @@ export class ResponseController {
     })
     file: File,
   ): Promise<SubmitApplicationResponse> {
+    if (!PermissionsService.canSubmitApplications(user)) {
+      throw new ForbiddenError();
+    }
     const response = await this.responseService.submitUserApplication(
       user,
       application,
@@ -79,6 +84,9 @@ export class ResponseController {
     })
     file?: File,
   ): Promise<SubmitApplicationResponse> {
+    if (!PermissionsService.canSubmitApplications(user)) {
+      throw new ForbiddenError();
+    }
     const response = await this.responseService.updateUserApplication(
       user,
       application,
@@ -92,6 +100,9 @@ export class ResponseController {
   async deleteApplication(
     @AuthenticatedUser() user: UserModel,
   ): Promise<DeleteApplicationResponse> {
+    if (!PermissionsService.canSubmitApplications(user)) {
+      throw new ForbiddenError();
+    }
     await this.responseService.deleteUserApplication(user);
     return { error: null };
   }
